quantlib/requirements.txt
numpy==1.21.2
python-decouple==3.4
redis==3.5.3
